dms_upload_api/requirements.txt
attrs==22.2.0
boto3==1.26.102
botocore==1.29.102
click==8.1.3
flasgger==0.9.5
Flask==2.2.3
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
